pyproject.toml
[project]
name = "combibench"
version = "0.1.0"
description = "Benchmarking LLM Capability for Combinatorial Mathematics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.18",
    "anthropic>=0.50.0",
    "datasets>=3.5.0",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "openai>=1.76.0",
    "pyjson5>=1.6.8",
    "strenum>=0.4.15",
    "tenacity>=9.1.2",
    "together>=1.5.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["evaluation"]

[project.optional-dependencies]
dev = ["pytest"]
